migrated sms to it's own brainmeat. replaced multiprocess dealio for incoming message checks with addlive()
