+++ brainmeats/nonsense.py
@@ -1,51 +1,21 @@
 import simplejson
 import urllib
 import urllib2
-import twilio
 import time
-import multiprocessing
 import os
 
 from autonomic import axon, category, help, Dendrite
-from settings import STORAGE, ACROLIB, LOGDIR, SHORTENER, DISTASTE, NICK, SMS_LOCKFILE
-from secrets import FML_API, TWILIO_SID, TWILIO_TOKEN, TWILIO_NUMBER 
+from settings import STORAGE, ACROLIB, LOGDIR, SHORTENER, DISTASTE, NICK
+from secrets import FML_API 
 from util import colorize
 from random import choice
 from xml.dom import minidom as dom
 from twilio.rest import TwilioRestClient
     
-def smsworker(self):
-    incoming = []
-    i = 0
-    n = 0 
-    
-    while True:
-        try:
-            client = TwilioRestClient(TWILIO_SID, TWILIO_TOKEN)
-            messages = client.sms.messages.list(to="+16468635380")
-        except:
-            n += 1
-
-        for item in messages:
-            message = item.from_ + ": " + item.body
-            if message not in incoming:
-                incoming.append(str(message))
-                if i > 0: 
-                    self.chat(message) 
-        
-        i += 1
-        time.sleep(10)
-    return
-
 @category("nonsense")
 class Nonsense(Dendrite):
     def __init__(self, cortex):
         super(Nonsense, self).__init__(cortex)
-       
-        if os.path.exists(SMS_LOCKFILE) == False:
-            open(SMS_LOCKFILE, 'a').close()
-            p = multiprocessing.Process(target=smsworker, args=(self,))
-            p.start()
 
     @axon
     @help("<generate bullshit>")
@@ -112,48 +82,6 @@ class Nonsense(Dendrite):
                 self.chat("No results. Or done broken.")
             else:
                 self.chat("Done broke")
-            return
-    
-    @axon
-    @help("[NUMBER] [MESSAGE] <send an sms message to unsuspecting victim>")
-    def sms(self):
-      
-        if not self.values:
-            self.chat("-sms <number> <message>")
-            return
-
-        to = self.values[0]
-        msg = " ".join(self.values[1:])
-
-        try:
-            client = TwilioRestClient(TWILIO_SID, TWILIO_TOKEN)
-            message = client.sms.messages.create(
-                body=msg,
-                to=to,
-                from_=TWILIO_NUMBER
-            )
-            self.chat("Message sent: " + message.sid)
-        except:
-            self.chat("Done broke")
-            return
-        
-    @axon
-    @help("[FROM NUMBER] <check for sms replies to messages previously sent to unsuspecting victims>")
-    def smsmsgs(self):
-        
-        if not self.values:
-            self.chat("What number?")
-            return
-
-        num = self.values[0]
-
-        try:
-            client = TwilioRestClient(TWILIO_SID, TWILIO_TOKEN)
-            messages = client.sms.messages.list(to="+16468635380",from_=num)
-            for item in messages:
-                self.chat(item.from_ + ": " + item.body) 
-        except:
-            self.chat("Done broke")
             return
 
     @axon
